scr/functions/trigger-dataproc/requirements.txt
google-cloud-dataproc==5.*
google-cloud-storage==2.*